Move error into its own module
