+++ server/server.rs
@@ -5,19 +5,15 @@
  */
 
 use std::{
-    fs, io,
+    fs,
     net::SocketAddr,
     path::{Path, PathBuf},
-    sync::{
-        atomic::{AtomicBool, Ordering},
-        Arc,
-    },
+    sync::Arc,
 };
 
 use concurrency::IntervalRunner;
-use database::{database_manager::DatabaseManager, DatabaseOpenError};
+use database::database_manager::DatabaseManager;
 use diagnostics::{diagnostics_manager::DiagnosticsManager, Diagnostics};
-use error::typedb_error;
 use rand::seq::SliceRandom;
 use resource::constants::server::{
     DATABASE_METRICS_UPDATE_INTERVAL, GRPC_CONNECTION_KEEPALIVE, SERVER_ID_ALPHABET, SERVER_ID_FILE_NAME,
@@ -35,6 +31,7 @@ use crate::{
     service::typedb_service::TypeDBService,
 };
 use resource::constants::server::ASCII_LOGO;
+use crate::error::ServerOpenError;
 
 #[derive(Debug)]
 pub struct Server {
@@ -54,7 +51,6 @@ pub struct Server {
 }
 
 impl Server {
-    // TODO: passing `deployment_id` as an arg to simply override it in Cloud. Consider refactoring
     pub async fn open(
         config: Config,
         distribution: &'static str,
@@ -306,23 +302,5 @@ impl Server {
 
     pub fn database_manager(&self) -> &DatabaseManager {
         self.typedb_service.as_ref().unwrap().database_manager()
-    }
-}
-
-typedb_error! {
-    pub ServerOpenError(component = "Server open", prefix = "SRO") {
-        NotADirectory(1, "Invalid path '{path}': not a directory.", path: String),
-        CouldNotReadServerIDFile(2, "Could not read data from server ID file '{path}'.", path: String, source: Arc<io::Error>),
-        CouldNotCreateServerIDFile(3, "Could not write data to server ID file '{path}'.", path: String, source: Arc<io::Error>),
-        CouldNotCreateDataDirectory(4, "Could not create data directory in '{path}'.", path: String, source: Arc<io::Error>),
-        InvalidServerID(5, "Server ID read from '{path}' is invalid. Delete the corrupted file and try again.", path: String),
-        DatabaseOpen(6, "Could not open database.", typedb_source: DatabaseOpenError),
-        Serve(7, "Could not serve on {address}.", address: SocketAddr, source: Arc<tonic::transport::Error>),
-        MissingTLSCertificate(8, "TLS certificate path must be specified when encryption is enabled."),
-        MissingTLSCertificateKey(9, "TLS certificate key path must be specified when encryption is enabled."),
-        CouldNotReadTLSCertificate(10, "Could not read TLS certificate from '{path}'.", path: String, source: Arc<io::Error>),
-        CouldNotReadTLSCertificateKey(11, "Could not read TLS certificate key from '{path}'.", path: String, source: Arc<io::Error>),
-        CouldNotReadRootCA(12, "Could not read root CA from '{path}'.", path: String, source: Arc<io::Error>),
-        TLSConfigError(13, "Failed to configure TLS.", source: Arc<tonic::transport::Error>),
     }
 }
